refactor the crashreporter module
Renamed the file from report to crashreporter and the modified
the methods of the class.

there are no side effects associated with these changes.
